update: project revamp
simplified the project structure to make it cleaner and simple.
