+++ config/app.go
@@ -1,22 +1,18 @@
 package config
 
 import (
-	"log"
 	"os"
 
-	"github.com/htanmo/todoapi/internal/api/v1/models"
+	"github.com/htanmo/todoapi/models"
 	"github.com/joho/godotenv"
 )
 
 func Load() {
-	err := godotenv.Load()
-	if err != nil {
-		log.Fatal("Failed to load .env file.")
-	}
+	godotenv.Load()
 
 	dbURI := os.Getenv("DB_URI")
 
-	// default db uri
+	// Default db uri
 	if dbURI == "" {
 		dbURI = "todo.db"
 	}
