+++ go.mod
@@ -5,12 +5,13 @@ go 1.22.3
 require (
 	github.com/joho/godotenv v1.5.1
 	github.com/rs/cors v1.11.0
-	gorm.io/driver/sqlite v1.5.5
-	gorm.io/gorm v1.25.10
+	gorm.io/driver/sqlite v1.5.6
+	gorm.io/gorm v1.25.11
 )
 
 require (
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.5 // indirect
 	github.com/mattn/go-sqlite3 v1.14.22 // indirect
+	golang.org/x/text v0.16.0 // indirect
 )
